script/requirements.txt
aioredis==1.2.0
asyncio==3.4.3
pytelegraf==0.3.3
redis==3.2.1
